Moved DayCard.js out of DayCardList.js.
